+++ src/main/java/org/example/expert/domain/manager/controller/ManagerController.java
@@ -3,7 +3,7 @@ package org.example.expert.domain.manager.controller;
 import io.jsonwebtoken.Claims;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.example.expert.config.JwtUtil;
+import org.example.expert.global.auth.JwtUtil;
 import org.example.expert.domain.common.annotation.Auth;
 import org.example.expert.domain.common.dto.AuthUser;
 import org.example.expert.domain.manager.dto.request.ManagerSaveRequest;
