+++ src/main/java/org/example/expert/global/auth/AuthUserArgumentResolver.java
@@ -1,4 +1,4 @@
-package org.example.expert.config;
+package org.example.expert.global.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.example.expert.domain.auth.exception.AuthException;
