+++ src/main/java/org/example/expert/global/auth/JwtFilter.java
@@ -1,4 +1,4 @@
-package org.example.expert.config;
+package org.example.expert.global.auth;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
