+++ src/main/java/org/example/expert/global/common/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package org.example.expert.config;
+package org.example.expert.global.common;
 
 import org.example.expert.domain.auth.exception.AuthException;
 import org.example.expert.domain.common.exception.InvalidRequestException;
