+++ src/main/java/org/example/expert/global/common/PasswordEncoder.java
@@ -1,4 +1,4 @@
-package org.example.expert.config;
+package org.example.expert.global.common;
 
 import at.favre.lib.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Component;
