+++ src/main/java/org/example/expert/global/config/FilterConfig.java
@@ -1,6 +1,8 @@
-package org.example.expert.config;
+package org.example.expert.global.config;
 
 import lombok.RequiredArgsConstructor;
+import org.example.expert.global.auth.JwtFilter;
+import org.example.expert.global.auth.JwtUtil;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
