+++ src/main/java/org/example/expert/web/weather/WeatherClient.java
@@ -1,6 +1,6 @@
-package org.example.expert.client;
+package org.example.expert.web.weather;
 
-import org.example.expert.client.dto.WeatherDto;
+import org.example.expert.web.weather.dto.WeatherDto;
 import org.example.expert.domain.common.exception.ServerException;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.http.HttpStatus;
