+++ src/test/java/org/example/expert/config/PasswordEncoderTest.java
@@ -1,5 +1,6 @@
 package org.example.expert.config;
 
+import org.example.expert.global.common.PasswordEncoder;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
